+++ midigen/models/__init__.py
@@ -1,4 +1,4 @@
-from .gpt1.gpt1 import MusicTransformer
+from .musictransformer.musictransformer import MusicTransformer
 from .gpt2.gpt2 import GPT
 
 __all__ = ['MusicTransformer', 'GPT']
